Removes outdated README and updates connection script
Removes the infra_demo README file, which is no longer relevant to the current setup.

Updates the connect.sh script with a new admin password and modifies resource group names for consistency in cluster setup.

Adds an entry to the .gitignore file for a specific PEM file to prevent it from being tracked.